Change :pick and :set to use bottom up stack indexing
This makes it much easier to keep track of where variables on the
stack are located as well as handling stack frames. With bottom up
indexing a variable (or frame) at index 0 always stays at index 0.
The top down indexing that :pick and :set used before was fiddly
because the index of a variable at index 0 changes index to 1 when
a value is pushed above it.