Disable TCO for tail calls inside let
Doing TCO for these calls is way harder than I first thought and
I won't be doing it in this project.

Why it's hard? The let has pushed a new stack frame that we need to
take into account. There is cleanup that needs to be done when exiting
the let. This cleanup can be moved to before making the call, but this
is way to tricky to be implemented without horrendous stack fiddling
or perhaps new instructions in the VM. Supporting this is for the next
project. Perhaps.